Revert "Cliente de velda"
