--- kickstart-main/README.md
@@ -1,2 +1,0 @@
-# kickstart
-template
