--- kickstart-main/src/co/edu/unbosque/controller/AplMain.java
@@ -1,11 +1,0 @@
-
-package co.edu.unbosque.controller;
-
-public class AplMain {
-
-	public static void main(String[] args) {
-		Controlador cont = new Controlador();
-
-	}
-
-}
